+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- CLI refactored to not using `process.exit()`
+
+### Added
+- test coverage: CLI
 
 <a id="latest"></a>
 ## [1.1.0-beta.5] - 2017-09-01